Handle mandatory entry point (mainpage and favicon).
Favicon is just a redirect to the real entry.

Main page need to be set in the header to the index of the mainpage.
Hovewer, it may break as the main page url ("/") can be different from
the real page address ("/foo/real/mainpage.h") and break links.

So client need to do a redirect to the real page. Let's integrate that
in the zim file itself by creating a redirect to the real mainpage.